base denominator increased to 10mil

+++ contracts/Core/RewardManager.sol
@@ -9,6 +9,7 @@ import "./interface/ICollectionManager.sol";
 import "../Initializable.sol";
 import "./storage/Constants.sol";
 import "./parameters/child/RewardManagerParams.sol";
+import "hardhat/console.sol";
 
 /// @title StakeManager
 /// @notice StakeManager handles stake, unstake, withdraw, reward, functions
@@ -126,7 +127,7 @@ contract RewardManager is Initializable, Constants, RewardManagerParams, IReward
             if (medianLastEpoch == 0) continue;
             uint64 prod = age * voteValueLastEpoch;
             // slither-disable-next-line calls-loop
-            uint16 tolerance = collectionManager.getCollectionTolerance(i);
+            uint32 tolerance = collectionManager.getCollectionTolerance(i);
             tolerance = tolerance <= maxTolerance ? tolerance : maxTolerance;
             uint64 maxVoteTolerance = medianLastEpoch + ((medianLastEpoch * tolerance) / BASE_DENOMINATOR);
             uint64 minVoteTolerance = medianLastEpoch - ((medianLastEpoch * tolerance) / BASE_DENOMINATOR);
